+++ src/spork.h
@@ -10,12 +10,12 @@
 #include "hash.h"
 #include "key.h"
 #include "main.h"
+#include "messagesigner.h"
 #include "net.h"
 #include "sporkid.h"
 #include "sync.h"
 #include "util.h"
 
-#include "obfuscation.h"
 #include "protocol.h"