+++ src/util.h
@@ -44,7 +44,6 @@ extern int64_t enforceMasternodePaymentsTime;
 extern std::string strMasterNodeAddr;
 extern int keysLoaded;
 extern bool fSucessfullyLoaded;
-extern std::vector<int64_t> obfuScationDenominations;
 extern std::string strBudgetMode;
 extern std::atomic<uint32_t> logCategories;
 extern std::map<std::string, std::string> mapArgs;